Rename code files for clarity
