+++ src/code/pauseHandler.js
